+++ loadouts/un_mo.hpp
@@ -1,0 +1,521 @@
+// in case of fire, tell Bear
+
+class baseMan {// Weaponless baseclass
+    displayName = "Unarmed";
+    // All randomized.
+    uniform[] = {"rhsgref_uniform_ttsko_mountain"};
+    vest[] = {};
+    backpack[] = {};
+    headgear[] = {};
+    goggles[] = {"default"};
+    hmd[] = {};
+    // Leave empty to remove all. "Default" > leave original item.
+
+    // All randomized
+    primaryWeapon[] = {};
+    scope[] = {};
+    bipod[] = {};
+    attachment[] = {};
+    silencer[] = {};
+    // Leave empty to remove all. "Default" for primaryWeapon > leave original weapon.
+
+    // Only *Weapons[] arrays are randomized
+    secondaryWeapon[] = {};
+    secondaryAttachments[] = {};
+    sidearmWeapon[] = {};
+    sidearmAttachments[] = {};
+    // Leave empty to remove all. "Default" for secondaryWeapon or sidearmWeapon > leave original weapon.
+
+    // These are added to the uniform or vest
+    magazines[] = {};
+    items[] = {"FirstAidKit"};
+    // These are added directly into their respective slots
+    linkedItems[] = {
+        "ItemMap",
+        "ItemCompass",
+        "ItemWatch"
+    };
+
+    // These are put into the backpack
+    backpackItems[] = {};
+
+    // This is executed after unit init is complete. argument: _this = _unit.
+    code = "";
+};
+class r : baseMan
+{
+    displayName = "Rifleman";
+    headgear[] = {"rhsgref_ssh68_un"};
+    vest[] = {"rhsgref_6b23_ttsko_mountain_rifleman"};
+    backpack[] = {"B_FieldPack_cbr"};
+    primaryWeapon[] = {"rhs_weap_m70ab2", "rhs_weap_m70ab2_fold"};
+    magazines[] =
+    {
+        LIST_9("rhs_30Rnd_762x39mm"),
+        LIST_2("rhs_30Rnd_762x39mm_tracer"),
+        LIST_2("rhs_mag_rgd5"),
+        LIST_2("rhs_mag_nspd")
+    };
+    items[] =
+    {
+        LIST_2("FirstAidKit")
+    };
+};
+class g : r
+{
+    displayName = "Grenadier";
+    primaryWeapon[] = {"rhs_weap_akms_gp25"};
+    magazines[] +=
+    {
+        LIST_8("rhs_VOG25"),
+        LIST_4("rhs_GRD40_White")
+    };
+};
+class car : r
+{
+    displayName = "Carabinier";
+};
+class m : car 
+{
+    displayName = "Medic";
+    vest[] = {"rhsgref_6b23_ttsko_mountain_medic"};
+    backpackItems[] = {
+        "Medikit",
+        LIST_7("FirstAidKit"),
+        LIST_2("rhs_mag_rdg2_white")
+    };
+};
+class smg : r
+{
+    displayName = "Submachinegunner";
+    primaryWeapon[] = {"rhs_weap_m92", "rhs_weap_m92_fold"};
+    magazines[] =
+    {
+        LIST_6("rhs_30Rnd_762x39mm"),
+        "rhs_mag_rgd5",
+        LIST_2("rhs_mag_rdg2_white")
+    };
+};
+class ftl : g
+{
+    displayName = "Fireteam Leader";
+    magazines[] +=
+    {
+        LIST_2("rhs_GRD40_Green"),
+        LIST_2("rhs_GRD40_Red"),
+        LIST_2("SmokeShellGreen")
+    };
+    linkedItems[] += {"Binocular","ItemGPS"};
+};
+class sl : ftl
+{
+    displayName = "Squad Leader";
+    vest[] = {"rhsgref_6b23_ttsko_mountain_officer"};
+    headgear[] = {"rhsgref_un_beret"};
+    sidearmWeapon[] = {"rhs_weap_makarov_pm"};
+    magazines[] += {
+        LIST_2("rhs_mag_9x18_8_57N181S")
+    };
+    linkedItems[] =
+    {
+        "ItemMap",
+        "ItemCompass",
+        "ItemWatch",
+        "Binocular",
+        "ItemGPS"
+    };
+};
+class co : sl
+{
+    displayName = "Platoon Leader";
+    radios[] = {"ACRE_PRC117F"};
+    backpackItems[] = {};
+};
+class fac : co
+{
+    displayName = "Forward Air Controller";
+    backpackItems[] = {};
+    radios[] = {"ACRE_PRC117F"};
+    linkedItems[] = {
+        "ItemMap",
+        "ItemCompass",
+        "ItemWatch",
+        "ItemGPS",
+        "Binocular"
+    };
+    items[] = {
+        LIST_2("FirstAidKit")
+    };
+};
+class ar : r
+{
+    displayName = "Automatic Rifleman";
+    vest[] = {"rhsgref_6b23_ttsko_mountain"};
+    primaryWeapon[] = {"rhs_weap_pkm"};
+    bipod[] = {};
+    silencer[] = {};
+    sidearmWeapon[] = {};
+    magazines[] =
+    {
+        LIST_4("rhs_100Rnd_762x54mmR"),
+        "rhs_mag_rgd5",
+        "rhs_mag_rdg2_white"
+    };
+};
+class aar : r
+{
+    displayName = "Assistant Automatic Rifleman";
+    backpackItems[] =
+    {
+        LIST_4("rhs_100Rnd_762x54mmR")
+    };
+    linkedItems[] += {"Binocular"};
+};
+class rat : car
+{
+    displayName = "Rifleman (AT)";
+    secondaryWeapon[] = {"rhs_weap_rpg26"};
+};
+class dm : r
+{
+    displayName = "Designated Marksman";
+    primaryWeapon[] = {"rhs_weap_svdp"};
+    silencer[] = {};
+    scope[] = {"rhs_acc_pso1m2"};
+    bipod[] = {};
+    magazines[] =
+    {
+        LIST_12("rhs_10Rnd_762x54mmR_7N1"),
+        "rhs_mag_rgd5",
+        LIST_2("rhs_mag_rdg2_white")
+    };
+};
+class mmgg : ar
+{
+    displayName = "MMG Gunner";
+    primaryWeapon[] = {"rhs_weap_pkp"};
+    sidearmWeapon[] = {};
+    scope[] = {"rhs_acc_pkas"};
+    magazines[] =
+    {
+        LIST_4("rhs_100Rnd_762x54mmR"),
+        LIST_2("rhs_100Rnd_762x54mmR"),
+        "rhs_mag_rgd5",
+        "rhs_mag_rdg2_white"
+    };
+};
+class mmgac : r
+{
+    displayName = "MMG Ammo Carrier";
+    backpackItems[] =
+    {
+        LIST_4("rhs_100Rnd_762x54mmR")
+    };
+};
+class mmgag : aar
+{
+    displayName = "MMG Assistant Gunner";
+    linkedItems[] =
+    {
+        "ItemMap",
+        "ItemCompass",
+        "ItemWatch",
+        "Binocular",
+        "ItemGPS"
+    };
+    backpackItems[] =
+    {
+        LIST_4("rhs_100Rnd_762x54mmR")
+    };
+};
+class hmgg : car
+{
+    displayName = "HMG Gunner";
+    backPack[] = {"RHS_Kord_Gun_Bag"};
+
+};
+class hmgac : r
+{
+    displayName = "HMG Ammo Carrier";
+    backPack[] = {"RHS_Kord_Gun_Bag"};
+};
+class hmgag : car
+{
+    displayName = "HMG Assistant Gunner";
+    backPack[] = {"RHS_Kord_Tripod_Bag"};
+    linkedItems[] =
+    {
+        "ItemMap",
+        "ItemCompass",
+        "ItemWatch",
+        "Binocular",
+        "ItemGPS"
+    };
+};
+class matg : car
+{
+    displayName = "MAT Gunner";
+    secondaryWeapon[] = {"rhs_weap_rpg7"};
+    secondaryAttachments[] = {"rhs_acc_pgo7v"};
+    backpack[] = {"rhs_rpg"};
+    magazines[] +=
+    {
+        "rhs_rpg7_PG7VR_mag",
+        LIST_2("rhs_rpg7_PG7VL_mag")
+    };
+};
+class matac : r
+{
+    displayName = "MAT Ammo Carrier";
+    backpack[] = {"rhs_rpg"};
+    backpackItems[] =
+    {
+        LIST_2("rhs_rpg7_PG7VR_mag"),
+        "rhs_rpg7_OG7V_mag"
+    };
+};
+class matag : car
+{
+    displayName = "MAT Assistant Gunner";
+    backpack[] = {"rhs_rpg"};
+    linkedItems[] =
+    {
+        "ItemMap",
+        "ItemCompass",
+        "ItemWatch",
+        "Binocular",
+        "ItemGPS"
+    };
+    backpackItems[] =
+    {
+        "rhs_rpg7_PG7VL_mag",
+        "rhs_rpg7_TBG7V_mag",
+        "rhs_rpg7_OG7V_mag"
+    };
+};
+class hatg : car
+{
+    displayName = "HAT Gunner";
+    backpack[] = {"RHS_SPG9_Tripod_Bag"};
+    secondaryWeapon[] = {};
+    backpackItems[] = {};
+};
+class hatac : r
+{
+    displayName = "HAT Ammo Carrier";
+    backpack[] = {"RHS_SPG9_Gun_Bag"};
+    secondaryWeapon[] = {};
+    backpackItems[] = {};
+};
+class hatag : car
+{
+    displayName = "HAT Assistant Gunner";
+    secondaryWeapon[] = {};
+    linkedItems[] =
+    {
+        "ItemMap",
+        "ItemCompass",
+        "ItemWatch",
+        "Binocular",
+        "ItemGPS"
+    };
+};
+class mtrg : car
+{
+    displayName = "Mortar Gunner";
+    backPack[] = {"RHS_Podnos_Gun_Bag"};
+    linkedItems[] =
+    {
+        "ItemMap",
+        "ItemCompass",
+        "ItemWatch",
+        "ItemGPS"
+    };
+};
+class mtrac : r
+{
+    displayName = "Mortar Ammo Carrier";
+    backPack[] = {"RHS_Podnos_Gun_Bag"};
+};
+class mtrag : car
+{
+    displayName = "Mortar Assistant Gunner";
+    backPack[] = {"RHS_Podnos_Bipod_Bag"};
+    linkedItems[] =
+    {
+        "ItemMap",
+        "ItemCompass",
+        "ItemWatch",
+        "Binocular",
+        "ItemGPS"
+    };
+};
+class samg : car
+{
+    displayName = "AA Missile Specialist";
+    secondaryWeapon[] = {"rhs_weap_igla"};
+    backpack[] = {"rhs_rpg"};
+    magazines[] +=
+    {
+        LIST_2("rhs_mag_9k38_rocket")
+    };
+};
+class samag : car
+{
+    displayName = "AA Assistant Missile Specialist";
+    backpack[] = {"rhs_rpg"};
+    linkedItems[] =
+    {
+        "ItemMap",
+        "ItemCompass",
+        "ItemWatch",
+        "Binocular",
+        "ItemGPS"
+    };
+    backpackItems[] =
+    {
+        LIST_2("rhs_mag_9k38_rocket")
+    };
+};
+class sn : r
+{
+    displayName = "Sniper";
+    vest[] = {"rhsgref_6b23_ttsko_mountain_sniper"};
+    headgear[] = {};
+    goggles[] = {};
+    primaryWeapon[] = {"rhs_weap_svdp"};
+    scope[] = {"rhs_acc_pso1m2"};
+    bipod[] = {};
+    sidearmWeapon[] = {"rhs_weap_pya"};
+    magazines[] =
+    {
+        LIST_13("rhs_10Rnd_762x54mmR_7N1"),
+        "rhs_mag_rgd5",
+        LIST_3("rhs_mag_9x19_17")
+    };
+    backpack[] = {};
+    linkedItems[] += {"Binocular","ItemGPS"};
+    items[] += {"rhsgref_un_beret"};
+};
+class sp : sn
+{
+    displayName = "Spotter";
+    scope[] = {"rhs_acc_1p78"};
+    silencer[] = {"rhs_acc_dtk"};
+    primaryWeapon[] = {"rhs_weap_m76"};
+    magazines[] =
+    {
+        LIST_11("rhsgref_10Rnd_792x57_m76"),
+        LIST_2("rhs_mag_rgd5"),
+        LIST_2("rhs_mag_rdg2_white"),
+        LIST_4("rhs_mag_9x19_17")
+    };
+};
+class vc : smg
+{
+    displayName = "Vehicle Commander";
+    radios[] = {"ACRE_PRC117F"};
+    vest[] = {"rhsgref_6b23_ttsko_mountain_nco"};
+    headgear[] = {"rhs_tsh4"};
+    goggles[] = {};
+    linkedItems[] += {"Binocular","ItemGPS"};
+};
+class vd : smg
+{
+    displayName = "Vehicle Driver";
+    vest[] = {"rhsgref_6b23_ttsko_mountain"};
+    backpack[] = {"rhs_assault_umbts_engineer_empty"};
+    headgear[] = {"rhs_tsh4"};
+    goggles[] = {};
+    linkedItems[] += {"ItemGPS"};
+    backpackItems[] = {"ToolKit"};
+};
+class vg : vd
+{
+    displayName = "Vehicle Gunner";
+    backpack[] = {};
+    backpackItems[] = {};
+};
+class pp : smg
+{
+    displayName = "Helicopter Pilot";
+    uniform[] = {"rhs_uniform_df15"};
+    radios[] = {"ACRE_PRC117F"};
+    headgear[] = {"rhs_zsh7a_mike"};
+    goggles[] = {};
+    linkedItems[] += {"ItemGPS"};
+    magazines[] += {
+        LIST_2("SmokeShellGreen")
+    };
+};
+class pcc : smg
+{
+    displayName = "Helicopter Crew Chief";
+    uniform[] = {"rhs_uniform_df15"};
+    headgear[] = {"rhs_zsh7a"};
+    goggles[] = {};
+    linkedItems[] += {"ItemGPS"};
+    backpackItems[] = {"ToolKit"};
+    magazines[] += {
+        LIST_2("SmokeShellGreen")
+    };
+};
+class pc : pcc
+{
+    displayName = "Helicopter Crew";
+    backpack[] = {};
+    backpackItems[] = {};
+};
+class jp : baseMan
+{
+    displayName = "Jet pilot";
+    uniform[] = {"rhs_uniform_df15"};
+    radios[] = {"ACRE_PRC117F"};
+    headgear[] = {"rhs_zsh7a"};
+    goggles[] = {};
+    sidearmWeapon[] = {"rhs_weap_pya"};
+    magazines[] =
+    {
+        LIST_4("rhs_mag_9x19_17")
+    };
+    items[] =
+    {
+        LIST_2("FirstAidKit")
+    };
+    linkedItems[] = {"ItemMap","ItemGPS","ItemCompass","ItemWatch"};
+};
+class eng : car
+{
+    displayName = "Combat Engineer (Explosives)";
+    vest[] = {"rhs_6b23_digi_vydra_3m"};
+    backpack[] = {"rhs_assault_umbts_engineer_empty"};
+    backpackItems[] = {
+        "MineDetector",
+        "ToolKit",
+        LIST_1("SatchelCharge_Remote_Mag"),
+        LIST_2("DemoCharge_Remote_Mag")
+    };
+    // TO REPLACE: items[] += {"ACE_M26_Clacker","ACE_DefusalKit"};
+};
+class engm : car
+{
+    displayName = "Combat Engineer (Mines)";
+    vest[] = {"rhs_6b23_digi_vydra_3m"};
+    backpack[] = {"B_Carryall_oli"};
+    items[] +=
+    {
+        // TO REPLACE: "ACE_DefusalKit"
+    };
+    backpackItems[] = {
+        "MineDetector",
+        "ToolKit",
+        LIST_3("rhs_mine_pmn2_mag"),
+        LIST_3("rhs_mine_tm62m_mag")
+    };
+};
+class UAV : car
+{
+    displayName = "UAV Operator";
+    backpack[] = {};
+    linkedItems[] += {"O_UavTerminal"};
+};
